refactor to check if collections exist before dropping

+++ lib/db/seeds/seedLocalDb.ts
@@ -1,25 +1,52 @@
 import { usersData } from '../data/usersData';
+import { eventsData } from '../data/eventsData';
 import { User, Event } from '../../Schemas/Schemas';
 import mongoose from 'mongoose';
 import { dbURL } from '../connection';
 
 // Seeding local database (connection required)
 const seedLocalDb = () => {
-    mongoose.connect(dbURL);
+    const connect = mongoose.createConnection(dbURL);
 
-    User.collection.drop();
-    //commented out event as doesn't exist yet
-    // Event.collection.drop();
-
-    User.insertMany(usersData)
-        .then((user: any) => {
-            console.log(`${user.length} users seeded`);
-        })
-        .catch((err: any) => {
-            console.log(err);
-        })
-        .finally(() => {
-            mongoose.connection.close();
-        });
+    connect.on('open', () => {
+        connect.db
+            .listCollections({ name: 'users' })
+            .toArray((err, collection) => {
+                if (err) {
+                    console.log(err);
+                    return;
+                }
+                if (collection!.length > 0) {
+                    User.collection.drop();
+                    console.log('Users collection dropped');
+                }
+                User.insertMany(usersData)
+                    .then((user) => {
+                        console.log(`${user.length} users seeded`);
+                    })
+                    .catch((err) => {
+                        console.log(err);
+                    });
+            });
+        connect.db
+            .listCollections({ name: 'events' })
+            .toArray((err, collectionNames) => {
+                if (err) {
+                    console.log(err);
+                    return;
+                }
+                if (collectionNames!.length > 0) {
+                    Event.collection.drop();
+                    console.log('Events collection dropped');
+                }
+                Event.insertMany(eventsData)
+                    .then((events) => {
+                        console.log(`${events.length} events seeded`);
+                    })
+                    .catch((err) => {
+                        console.log(err);
+                    });
+            });
+    });
 };
 seedLocalDb();
